+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Features
 - Multiple solutions to autoclose tickets #19630
 - Allow anonymous requester #19798
+- See my group-assigned tickets only profile #20128
 
 ## 2.2.3 - 2024-01-15
 ### Bugfixes